+++ components/ImageVisualizer.tsx
@@ -14,7 +14,7 @@ import { extractStatusFromError } from '../utils/aiErrorUtils';
 import { dispatchAIRequest } from '../services/modelDispatcher';
 import { MINIMAL_MODEL_NAME, AUXILIARY_MODEL_NAME } from '../constants';
 import Button from './elements/Button';
-import { Icon } from './icons.tsx';
+import { Icon } from './elements/icons';
 
 const THEME_STYLE_PROMPTS: Record<string, string> = {
   dungeon: 'a dark, gritty medieval fantasy style, dungeons and dragons concept art',
