+++ components/InventoryDisplay.tsx
@@ -3,7 +3,7 @@
  * @description Shows the Player's items and handles interactions.
  */
 import { Item, KnownUse } from '../types';
-import { Icon } from './icons';
+import { Icon } from './elements/icons';
 import InventoryItem from './InventoryItem';
 import InventorySortControls from './InventorySortControls';
 import { useInventoryDisplay } from '../hooks/useInventoryDisplay';
